eliminate duplicated fixture AssertDataframesConfiguration (and a few more things)
