+++ source/geh_calculated_measurements/tests/capacity_settlement/job_tests/conftest.py
@@ -15,7 +15,7 @@ from tests.capacity_settlement.job_tests import TEST_FILES_FOLDER_PATH
 
 
 @pytest.fixture(scope="session")
-def seed_gold_table(spark: SparkSession) -> None:
+def gold_table_seeded(spark: SparkSession) -> None:
     create_database(spark, MeasurementsGoldDatabaseDefinition.DATABASE_NAME)
 
     create_table(
@@ -36,7 +36,7 @@ def seed_gold_table(spark: SparkSession) -> None:
 
 
 @pytest.fixture(scope="session")
-def create_calculated_measurements_table(spark: SparkSession, test_files_folder_path: str) -> None:
+def calculated_measurements_table_created(spark: SparkSession) -> None:
     create_database(spark, CalculatedMeasurementsInternalDatabaseDefinition.DATABASE_NAME)
 
     create_table(
