+++ source/geh_calculated_measurements/tests/data_products/conftest.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from typing import Generator
 
 import pytest
-from geh_common.testing.dataframes import AssertDataframesConfiguration, write_when_files_to_delta
+from geh_common.testing.dataframes import write_when_files_to_delta
 from geh_common.testing.scenario_testing import TestCase, TestCases, get_then_names
 from pyspark.sql import SparkSession
 
@@ -12,7 +12,6 @@ from geh_calculated_measurements.common.infrastructure import CalculatedMeasurem
 from geh_calculated_measurements.database_migrations.migrations_runner import migrate
 from geh_calculated_measurements.database_migrations.settings.catalog_settings import CatalogSettings
 from tests import PROJECT_ROOT
-from tests.testsession_configuration import TestSessionConfiguration
 
 
 @pytest.fixture(scope="module", autouse=True)
@@ -91,14 +90,3 @@ def test_cases(spark: SparkSession, request: pytest.FixtureRequest) -> TestCases
 
     # Return test cases
     return TestCases(test_cases)
-
-
-@pytest.fixture(scope="session")
-def assert_dataframes_configuration(
-    test_session_configuration: TestSessionConfiguration,
-) -> AssertDataframesConfiguration:
-    return AssertDataframesConfiguration(
-        show_actual_and_expected_count=test_session_configuration.scenario_tests.show_actual_and_expected_count,
-        show_actual_and_expected=test_session_configuration.scenario_tests.show_actual_and_expected,
-        show_columns_when_actual_and_expected_are_equal=test_session_configuration.scenario_tests.show_columns_when_actual_and_expected_are_equal,
-    )
